+++ src/main/java/strings/Task18.java
