+++ src/main/java/strings/Task19.java
