+++ src/main/java/strings/Task20.java
