+++ src/main/java/task25/interfaces/CardOperations.java
@@ -1,0 +1,6 @@
+package task25.interfaces;
+
+public interface CardOperations {
+    void getBalance();
+    int moneyAfterCash(int needMoney);
+}
