--- task25/CardOperations.java
@@ -1,7 +1,0 @@
-package task25;
-
-public interface CardOperations {
-
-    void getBalance();
-    int moneyAfterCash(int needMoney);
-}
